% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{afib}
\alias{afib}
\title{Atrial Fibrillation data}
\format{A dataset.}
\usage{
data(afib)
}
\description{
Data from the R package gemtc. Description and references will be added soon.
}
\examples{
data(afib)

}
\keyword{datasets}
