% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.model.fit.R
\name{assess.model.fit}
\alias{assess.model.fit}
\title{Assess Model Fit}
\usage{
assess.model.fit(jagsoutput, ...)
}
\arguments{
\item{jagsoutput}{Resulting dataset from running \code{nma.analysis()}.}

\item{...}{Graphical arguments such as main=, ylab=, and xlab= may be passed as in \code{plot()}. These arguments will only effect the
leverage plot.}
}
\value{
\code{DIC} - A number indicating the Deviance Information Criteria. A larger DIC is indicative of a worse model fit.

\code{leverage} - A vector indicating the leverage of each data point (study arm). Leverage is defined as the posterior mean of the
residual deviance minus the deviance at the posterior mean of the fitted values.

\code{w} - A vector with 1 value per study arm. The magnitude of \code{w} represents the data point's contribution to the
posterior mean deviance of the model. The sign indicates whether the data is being over or under estimated by the model.

\code{pmdev} -
}
\description{
Computes the Deviance Information Criteria (DIC) and produces a leverage plot (as in the NICE Technical Support Document 2)
for a given model. These can be used to assess and compare the fit of different models (i.e fixed vs random effects, consistency vs
inconsistency). \code{assess.model.fit} also produces a plot comparing the leverage of each data point against their contribution to
the total posterior deviance. Any point lying outside the purple dotted line is considered to be an "outlier";
contributing to the model's poor fit.
}
\examples{
#Compare fixed vs random effects via leverage plots and DIC
#fixed_effects_results are the outputs of nma.analysis() where the parameter effects="fixed"
#random_effects_results are the outputs of nma.analysis() where the parameter effects="random"

par(mfrow=c(1,2))
assess.model.fit(fixed_effects_results, main = "Fixed Effects Model" )
assess.model.fit(random_effects_results, main= "Random Effects Model")

}
