% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUGSnet-package.R
\docType{package}
\name{BUGSnet}
\alias{BUGSnet}
\alias{BUGSnet-package}
\title{BUGSnet: Bayesian network meta-analyses in compliance with best practice guidelines.}
\description{
BUGSnet (NETwork meta-analysis with BUGS) is a feature-rich R package to conduct Bayesian network meta-analyses in compliance with best practice guidelines and facilitate reporting.
Bayesian analyses are conducted with JAGS and BUGS code is automatically generated by the package based on the user's inputs.
Outputs include network plots, tables of network characteristics, league tables and league heat plots, SUCRA plots, rankograms, forest plots, leverage plots, traceplots, posterior mean deviance comparison plots.
}
