% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{contrast}
\alias{contrast}
\title{?? data}
\format{
A dataset.
}
\usage{
data(contrast)
}
\description{
Data from COvalence research in the form of contrast-based data
}
\examples{
data(contrast)

}
\keyword{datasets}
