% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.prep.R
\name{data.prep}
\alias{data.prep}
\title{Data preparation}
\usage{
data.prep(arm.data, varname.t, varname.s)
}
\arguments{
\item{arm.data}{Data with 1 row for each study arm.}

\item{varname.t}{A string indicating the name of the treatment variable.}

\item{varname.s}{A string indicating the name of the study variable.}
}
\value{
\code{arm.data} - A tibble containing the arm level study data

\code{treatments} - A list of all treatments in the network

\code{studies} - A list of all studies in the network

\code{n.arms} - A tibble containing the number of arms for each study

\code{varname.t} - A string containing the name of the treatment variable

\code{varname.s} - A string containing the name of the study variable
}
\description{
Puts data into appropriate form for subsequent analysis using BUGSnet.
}
\examples{
# Example 1
data(diabetes.sim)
diabetes.slr <- data.prep(arm.data = diabetes.sim, 
varname.t = "Treatment", 
varname.s = "Study")
diabetes.slr$arm.data
diabetes.slr$treatments
diabetes.slr$studies
}
\seealso{
\code{\link{net.plot}}, \code{\link{net.tab}}, \code{\link{data.plot}}, \code{\link{nma.model}}
}
