% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes data}
\format{A dataset.}
\usage{
data(diabetes)
}
\description{
Data from NICE TSD 2. Description and references will be added soon.
}
\examples{
data(diabetes)

}
\keyword{datasets}
