% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dietfat}
\alias{dietfat}
\title{Dietfat data}
\format{A dataset.}
\usage{
data(dietfat)
}
\description{
Data from NICE TSD 2. Description and references will be added soon.
}
\examples{
data(dietfat)

}
\keyword{datasets}
