% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.model.fit.R
\name{inconsistency.plot}
\alias{inconsistency.plot}
\title{Consistency vs Inconsistency plot}
\usage{
inconsistency.plot(consistency.model.fit, inconsistency.model.fit, ...)
}
\arguments{
\item{consistency.model.fit}{Results of \code{assess.model.fit()} of an consistency model.}

\item{inconsistency.model.fit}{Results of \code{assess.model.fit()} of an inconsistency model.}

\item{...}{Graphical arguments such as main=, ylab=, and xlab= may be passed as in \code{plot()}.}
}
\description{
Plots the posterior mean deviance of a consistency model vs an inconsistency model. This plot can help identify loops
where inconsistency is present. Ideally, both models will contribute approximately 1 to the posterior mean deviance.
}
\examples{
# Assess model fit for a both an inconsistency model and consistency model using assess.model.fit()
assess.consistency <- assess.model.fit(consistency_results)
assess.inconsistency <- assess.model.fit(inconsistency_results)

#Plot the results against each other to assess inconsistency
inconsistency.plot(consistency_results, inconsistency_results)
}
