% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.heat.plot.R
\name{league.heat.plot}
\alias{league.heat.plot}
\title{League Heat Plot}
\usage{
league.heat.plot(leaguetable, low.colour = "red", mid.colour = "white",
  high.colour = "springgreen4", midpoint = 0)
}
\arguments{
\item{leaguetable}{Results from \code{leaguetable()}. Be sure to set layout="long" from \code{leaguetable()} (see example).}

\item{low.colour}{A string indicating the colour of negative relative treatment effects (e.g relative risk of 0.5).}

\item{mid.colour}{A string indicating the colour of null relative treatment effects (e.g relative risk of ~1.0).}

\item{high.colour}{A string indicating the colour of high relative treatment effects (e.g relative risk of ~2.0).}

\item{midpoint}{Point indicating a null treatment effect. THIS NEEDS TO BE UPDATED. CURRENTLY ONLY WORKS FOR BINOMIAL VARIABLES.}
}
\description{
Creates and plots a heatmap for a league table produced by \code{leaguetable()}. The user inputs a colour indicating
a negative relative treatment effect, a null relative treatment effect, and a large relative treatment effect.
}
\examples{
#Make league table, be sure to set layout="long".
lt <- leaguetable(jagsoutput=nma.results, central.tdcy="median", layout = "long")

#make plot using default colours
league.heat.plot(leaguetable=lt)
}
