% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.table.R
\name{leaguetable}
\alias{leaguetable}
\title{League Table}
\usage{
leaguetable(jagsoutput, central.tdcy = "median", layout = "default")
}
\arguments{
\item{jagsoutput}{Results from running \code{nma.analysis()}.}

\item{central.tdcy}{The statistic that you want to use in order to measure relative effectiveness. The options are "mean" and "median".}

\item{layout}{Indicates the form of the results. The options are "default" and "long". Use "default" if you want an nxn matrix
comparing all n treatments. Use "long" if you want 1 line per treatment comparison. Note that if you want to use the function 
\code{league.heat.plot}, set this parameter to "long".}
}
\description{
Produces league table that contains all information about relative effectiveness and their 
uncertainty for all possible pairs of treatments.
}
\examples{

league_table <- leaguetable(jagsoutput=nma.results, central.tdcy="median", layout = "long")
}
\seealso{
\code{league.heat.plot()}
}
