% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.plot.R
\name{net.plot}
\alias{net.plot}
\title{Network Plot}
\usage{
net.plot(data, node.scale = 5, edge.scale = 2, flag = NULL,
  label.offset1 = 0, label.offset2 = 1, graph.scale = T,
  node.lab.cex = 1, edge.lab.cex = 1, node.colour = "#f69c54",
  edge.colour = "#4a5b71", edge.lab.colour = "blue",
  flag.edge.colour = "lightpink", ...)
}
\arguments{
\item{data}{An object produced by \code{data.prep()} containing the data.}

\item{node.scale}{Size of the nodes (default=5)}

\item{edge.scale}{Thickness of the edges (default=2)}

\item{flag}{Used to highlight direct comparisons to a particular treatment (optional.
Set this value to treatment of interest and it will highlight, in red, all of the edges
going into this treatment.}

\item{label.offset1}{Node label location (x-axis) relative to node. Default=0}

\item{label.offset2}{Node label location (y-axis) relative to node. Default=1}

\item{graph.scale}{bla}

\item{node.lab.cex}{bla}

\item{edge.lab.cex}{bla}

\item{node.colour}{bla}

\item{edge.colour}{bla}

\item{edge.lab.colour}{bla}

\item{flag.edge.colour}{bla}

\item{...}{bla}
}
\description{
Produces network plot where nodes represent treatments and edges represent direct
evidence (e.g an RCT) comparing treatments.
}
\examples{

data(diabetes.sim)

diabetes.slr <- data.prep(arm.data = diabetes.sim, 
varname.t = "Treatment", 
varname.s = "Study")

# use default settings
net.plot(diabetes.slr)

# Highlight all direct comparisons with Placebo. Adjust node and edge size, centre node labels
net.plot(diabetes.slr, 
node.scale=4, 
edge.scale=1.5, 
flag="Placebo", 
label.offset1=0, 
label.offset2=0)
}
\seealso{
\code{\link{data.prep}}
}
