% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.tab.R
\name{net.tab}
\alias{net.tab}
\title{Generate Network Characteristics}
\usage{
net.tab(data.nma, outcome, N, type.outcome, time)
}
\arguments{
\item{data.nma}{A data object produced by \code{data.prep()}}

\item{outcome}{A string indicating the name of your outcome variable}

\item{N}{A string indicating the name of the variable containing the number of participants in each arm}

\item{type.outcome}{A string. Options are: "binomial", "continuous", "rate" (e.g # of events and # person-years reported), 
"rate2" (e.g # events and followup time reported)}

\item{time}{A string required when type.outcome = "rate" or "rate2". Name of variable 
indicating person-time followup (e.g person years) or study followup time}
}
\value{
\code{network} - Table of network characteristics

\code{intervention} - Summary statistics broken down by treatment

\code{comparison} - Summary statistics broken down by treatment comparison
}
\description{
Generates tables of network characteristics
}
\examples{
network.charac(data.nma = data.prep(my.data), outcome = "n_died", N = "n", type.outcome="binomial", time = NULL)
network.charac(data.nma = data.prep(my.data), outcome = "y", N = "N", type.outcome="rate", time = "personYears")
}
