% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.plot.R
\name{network.plot}
\alias{network.plot}
\title{Network Plot}
\usage{
network.plot(slr, node.scale = 5, edge.scale = 2, flag = NULL,
  label.offset1 = 0, label.offset2 = 1)
}
\arguments{
\item{slr}{An slr data object produced by \code{data.slr()}}

\item{node.scale}{Size of the nodes (default=5)}

\item{edge.scale}{Thickness of the edges (default=2)}

\item{flag}{Used to highlight direct comparisons to a particular treatment (optional.
Set this value to treatment of interest and it will highlight, in red, all of the edges
going into this treatment.}

\item{label.offset1}{Node label location (x-axis) relative to node. Default=0}

\item{label.offset2}{Node label location (y-axis) relative to node. Default=1}
}
\description{
Produces network plot where nodes represent treatments and edges represent direct
evidence (e.g an RCT) comparing treatments.
}
\examples{
# use default settings
network.plot(my.slr)

# Highlight all direct comparisons with Placebo. Adjust node and edge size, centre node labels
network.plot(my.slr, node.scale=4, edge.scale=1.5, flag="Placebo", label.offset1=0, label.offset2=0)
}
