% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.analysis.R
\name{nma.analysis}
\alias{nma.analysis}
\title{Run NMA model}
\usage{
nma.analysis(bugs, monitor = c("d"), n.adapt, n.burnin, n.iter,
  thin = 1, n.chains = 3)
}
\arguments{
\item{bugs}{Object produced by running \code{nma.bugs}.}

\item{monitor}{A list of all variables that you would like to monitor. The default is simply the treatment
effect samples ("d"). But you may want to monitor the deviance ("dev") as well.}
}
\value{
\code{model} - A long character string containing the model that was run in \code{rjags}.

\code{bugsdata2}

\code{scale}

\code{trt.map.table} - Treatments mapped to integer numbers, used to run BUGS code.
}
\description{
Takes bugs code from an object produced by \code{nma.bugs} and runs model using \code{rjags}.
}
