% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.bugs.R
\name{nma.bugs}
\alias{nma.bugs}
\title{Make Bugs Model}
\usage{
nma.bugs(slr, outcome, N, baseline.name, type = NULL, sd = NULL,
  exposure.time = NULL, family, link, effects)
}
\arguments{
\item{slr}{An slr data object produced by \code{data.slr()}}

\item{outcome}{A string indicating the name of your outcome variable}

\item{N}{A string indicating the name of the variable containing the number of participants in each arm}

\item{baseline.name}{This is the drug that all treatments will be compared to in the network. This is often
a placebo or control drug of some kind.}

\item{type}{If type="inconsistency", an inconsistency model will be built. Otherwise a consistency model
will be built}

\item{sd}{A string (only required for continuous outcomes) indicating variable name
of the standard deviation of the outcome}

\item{exposure.time}{A string (only required for rate or rate2 outcomes) indicating the name of variable 
indicating person-time followup (e.g person years) or study followup time.}

\item{family}{A string indicating the family of the distribution of the outcome. Options are:
"binomial", "normal", "poisson".}

\item{link}{The link function for the nma model. Options are "logit", "log", "cloglog", "identity".}

\item{effects}{Options are "fixed" or "random".}
}
\value{
\code{model.str} - A long character string containing BUGS code that will be run in \code{rjags}.

\code{bugsdata2} - Data accompanying BUGS code.

\code{scale} - The scale of the outcome, based on the chosen family and link function
examples are "RR", "OR", "SMD", "HR"

\code{trt.map.table} - Treatments mapped to integer numbers, used to run BUGS code.
}
\description{
Creates bugs BUGS code which will can be ran through \code{nma.analysis()}.
}
\examples{
#Example 1
#Fixed effects, consistency model.
#Binomial family, log link. This implies that the relative risk (RR) will be the scale.

nma.bugs(slr = my.slr,
       outcome = "n_died", 
       N = "n",
       baseline.name = "plbo",
       family = "binomial",
       link = "log",
       effects = "fixed")
       
#Example 2
#Random Effects, inconsistency model
#Normal family, identity link.        
 
nma.bugs(slr = my.slr,
        outcome = "weight_change", 
        N = "n",
        baseline.name = "control",
        type = "inconsistency",
        sd = "std_dev",
        family = "normal",
        link = "identity",
        effects = "random")
        
# Example 3
# Random Effects, consistency model,
# Poisson family, cloglog link.

nma.bugs(slr = my.slr,
        outcome = "n_died", 
        N = "n",
        baseline.name = "plbo",
        exposure.time="personyears",
        family = "poisson",
        link = "cloglog",
        effects = "random")
}
