% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.forest.R
\name{nma.forest}
\alias{nma.forest}
\title{Forest plot}
\usage{
nma.forest(
  nma,
  comparator,
  central.tdcy = "median",
  order = NULL,
  log.scale = FALSE,
  lwd = 1,
  x.trans = NULL,
  cov.value = NULL
)
}
\arguments{
\item{nma}{A \code{BUGSnetRun} object produced by running \code{nma.run()}.}

\item{comparator}{The treatment to use as a comparator}

\item{central.tdcy}{The posterior statistic used in order to measure relative effectiveness. The options are "mean" and "median". Default is median.}

\item{order}{Optional. A vector of strings representing the order in which to display the treatments.}

\item{log.scale}{If TRUE, odds ratios, relative risk or hazard ratios are reported on the log scale. Default is FALSE.}

\item{lwd}{Line width relative to the default (default=1).}

\item{x.trans}{Optional. A string indicating a transformation to apply to the x-axis. Setting this parameter to "log" is useful when there are extreme values or to allow an easier interpretation of odds ratios and relative ratios (if e.g. treatment B is twice as far from the line y=1 then treatment A then it's OR/RR is twice that of treatment A.)}

\item{cov.value}{Must be specified for meta-regression. This is the value of the covariate for which to report the results.}
}
\value{
\code{forestplot} - A forest plot.
}
\description{
Produces a forest plot of point estimates and 95\% credible intervals obtained with the quantile method.
}
\examples{
data(diabetes.sim)

diabetes.slr <- data.prep(arm.data = diabetes.sim, 
varname.t = "Treatment", 
varname.s = "Study")

#Random effects, consistency model.
#Binomial family, cloglog link. This implies that the scale will be the Hazard Ratio.
diabetes.re.c <- nma.model(data = diabetes.slr,
       outcome = "diabetes", 
       N = "n",
       reference = "Placebo",
       family = "binomial",
       link = "cloglog",
       effects = "random",
       type="consistency",
       time="followup"
       )
 
diabetes.re.c.res <- nma.run(diabetes.re.c,
n.adapt=100,
n.burnin=0,
n.iter=100)
     
       
#make forest plot
nma.forest(nma = diabetes.re.c.res, comparator="Placebo")
}
\seealso{
\code{\link{nma.run}}, \code{\link{nma.league}}, \code{\link{nma.rank}}
}
