% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.league.R
\name{nma.league}
\alias{nma.league}
\title{League Table and Heat Plot}
\usage{
nma.league(jagsoutput, central.tdcy = "median", sucra.ranks = NULL,
  low.colour = "red", mid.colour = "white",
  high.colour = "springgreen4", midpoint = 0)
}
\arguments{
\item{jagsoutput}{Results from running \code{nma.analysis()}.}

\item{central.tdcy}{The statistic that you want to use in order to measure relative effectiveness. The options are "mean" and "median".}

\item{sucra.ranks}{A string of the ordered treatments for the heatmap.}

\item{low.colour}{A string indicating the colour of negative relative treatment effects for the heat plot (e.g relative risk of 0.5).}

\item{mid.colour}{A string indicating the colour of null relative treatment effects for the heat plot (e.g relative risk of ~1.0).}

\item{high.colour}{A string indicating the colour of high relative treatment effects for the heat plot (e.g relative risk of ~2.0).}

\item{midpoint}{Value indicating a null treatment effect for the heat plot. THIS NEEDS TO BE UPDATED. CURRENTLY ONLY WORKS FOR BINOMIAL VARIABLES.}
}
\value{
\code{table} - League table.

\code{plot} - League table heatplot, where a color scale is used to represent relative treatment effects.
}
\description{
Produces a league table and a league heat plot that contain all information about relative effectiveness and their 
uncertainty for all possible pairs of treatments.
}
\examples{

league_table <- leaguetable(jagsoutput=nma.results, central.tdcy="median")
}
