% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.model.contrast.R
\name{nma.model.contrast}
\alias{nma.model.contrast}
\title{Create Bugs Model for contrast-based data}
\usage{
nma.model.contrast(
  data_contrast = NULL,
  differences,
  se.diffs = NULL,
  var.ref = NULL,
  reference,
  type = "consistency",
  time = NULL,
  effects,
  scale,
  prior.mu = "DEFAULT",
  prior.d = "DEFAULT",
  prior.sigma = "DEFAULT"
)
}
\arguments{
\item{data_contrast}{A \code{BUGSnetData} object containing the data from contrast-based trials produced by \code{data.prep()}}

\item{differences}{A string indicating the name of the differences for contrast-based studies}

\item{var.ref}{A string (only required for networks with multi-arm trials) indicating the variable name of the variance of the reference treatment in each study}

\item{reference}{A string for the treatment that will be seen as the 'referent' comparator and labeled as treatment 1 in the BUGS code. This is often
a placebo or control drug of some kind.}

\item{type}{If type="inconsistency", an inconsistency model will be built. By default, type="consistency" and a consistency model is built.
will be built.}

\item{effects}{A string indicating the type of treatment effect relative to baseline. Options are "fixed" or "random".}

\item{scale}{A string indicating the scale of the data, such as "Mean Difference" or "Log-Odds Ratio",}

\item{prior.mu}{A string of BUGS code that defines priors on the baseline treatment effects. By default, independent normal priors are used with mean 0 and standard deviation 15u, where u is the largest maximum likelihood estimator in single trials \insertCite{@see @gemtc}{BUGSnet}.}

\item{prior.d}{A string of BUGS code that defines define priors on relative treatment effects. By default, independent normal priors are used with mean 0 and standard deviation 15u, where u is the largest maximum likelihood estimator in single trials \insertCite{@see @gemtc}{BUGSnet}.}

\item{prior.sigma}{A string of BUGS code that defines the prior on the variance of relative treatment effects. By default, a uniform distribution with range 0 to u is used, where u is the largest maximum likelihood estimator in single trials \insertCite{@see @gemtc}{BUGSnet}.}

\item{se.c}{A string indicating the variable name of the standard errors of the differences.}
}
\value{
\code{nma.model} returns an object of class \code{BUGSnetModel} which is a list containing the following components:

\code{bugs} - A long character string containing BUGS code that will be run in \code{jags}.

\code{data} - The data used in the BUGS code.

\code{scale} - The scale of the outcome, based on the chosen family and link function
examples are "Risk Ratio" (relative risk), "Odds Ratio", "Mean Difference", "Hazard Ratio"

\code{trt.key} - Treatments mapped to integer numbers, used to run BUGS code.

...
}
\description{
Creates BUGS code which can be ran through \code{nma.run()}.
}
\examples{


data(diabetes.sim)

diabetes.slr <- data.prep(arm.data = diabetes.sim, 
varname.t = "Treatment", 
varname.s = "Study")

#Random effects, consistency model.
#Binomial family, cloglog link. This implies that the scale will be the Hazard Ratio.
diabetes.re.c <- nma.model(data = diabetes.slr,
       outcome = "diabetes", 
       N = "n",
       reference = "Placebo",
       family = "binomial",
       link = "cloglog",
       effects = "random",
       type="consistency",
       time="followup"
       )
       
#Fixed effects, consistency model.
#Binomial family, cloglog link. This implies that the scale will be the Hazard Ratio.
diabetes.fe.c <- nma.model(data = diabetes.slr,
       outcome = "diabetes", 
       N = "n",
       reference = "Placebo",
       family = "binomial",
       link = "cloglog",
       effects = "fixed",
       type="consistency",
       time="followup"
       )
        
}
\references{
\insertRef{gemtc}{BUGSnet}

\insertRef{TSD3}{BUGSnet}
}
\seealso{
\code{\link{data.prep}}, \code{\link{nma.run}}
}
