% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.model.R
\name{nma.model}
\alias{nma.model}
\title{Create Bugs Model}
\usage{
nma.model(data, outcome, N, sd = NULL, reference, type = "consistency",
  time = NULL, family, link, effects, prior.mu = "DEFAULT",
  prior.d = "DEFAULT", prior.sigma = "DEFAULT", prior.beta = "",
  covariate = NULL)
}
\arguments{
\item{data}{A data object produced by \code{data.prep()}}

\item{outcome}{A string indicating the name of your outcome variable.}

\item{N}{A string indicating the name of the variable containing the number of participants in each arm}

\item{sd}{A string (only required for continuous outcomes) indicating variable name
of the standard deviation of the outcome}

\item{reference}{A string for the treatment that will be seen as the 'referent' comparator and labeled as treatment 1 in the BUGS code. This is often
a placebo or control drug of some kind.}

\item{type}{If type="inconsistency", an inconsistency model will be built. By default, type="consistency" and a consistency model is built.
will be built.}

\item{time}{A string (only required for binomial-cloglog or poisson-log models) indicating the name of variable 
indicating person-time followup (e.g person years) or study followup time.}

\item{family}{A string indicating the family of the distribution of the outcome. Options are:
"binomial", "normal", "poisson".}

\item{link}{The link function for the nma model. Options are "logit", "log", "cloglog", "identity".}

\item{effects}{A string indicating the type of treatment effect relative to baseline. Options are "fixed" or "random".}

\item{prior.mu}{A string of BUGS code that defines priors on the baseline treatment effects. By default, independent normal priors are used with mean 0 and standard deviation 15u, where u is the largest maximum likelihood estimator in single trials \insertCite{@see @gemtc}{BUGSnet}.}

\item{prior.d}{A string of BUGS code that defines define priors on relative treatment effects. By default, independent normal priors are used with mean 0 and standard deviation 15u, where u is the largest maximum likelihood estimator in single trials \insertCite{@see @gemtc}{BUGSnet}.}

\item{prior.sigma}{A string of BUGS code that defines the prior on the variance of relative treatment effects. By default, a uniform distribution with range 0 to u is used, where u is the largest maximum likelihood estimator in single trials \insertCite{@see @gemtc}{BUGSnet}.}

\item{prior.beta}{Optional string that defines the prior on the meta-regression coefficients. Options are "UNRELATED", "EXCHANGEABLE", "EQUAL" \insertCite{@see @TSD3}{BUGSnet} or a string of BUGS code.}

\item{covariate}{Optional string indicating the name of the variable in your data set that you would like to
adjust for via meta regression. By default, covariate=NULL and no covariate adjustment is applied.}
}
\value{
\code{model} - A long character string containing BUGS code that will be run in \code{rjags}.

\code{bugsdata2} - Data accompanying BUGS code.

\code{scale} - The scale of the outcome, based on the chosen family and link function
examples are "RR", "OR", "SMD", "HR"

\code{trt.key} - Treatments mapped to integer numbers, used to run BUGS code.
}
\description{
Creates BUGS code which can be ran through \code{nma.run()}.
}
\details{

}
\examples{
#Example 1
#Fixed effects, consistency model.
#Binomial family, log link. This implies that the relative risk (RR) will be the scale.

nma.model(data = my.slr,
       outcome = "n_died", 
       N = "n",
       reference = "plbo",
       family = "binomial",
       link = "log",
       effects = "fixed")
       
#Example 2
#Random Effects, inconsistency model
#Normal family, identity link.        
 
nma.model(data = my.slr,
        outcome = "weight_change", 
        N = "n",
        reference = "control",
        type = "inconsistency",
        sd = "std_dev",
        family = "normal",
        link = "identity",
        effects = "random")
        
# Example 3
# Random Effects, consistency model,
# Poisson family, cloglog link.

nma.model(data = my.slr,
        outcome = "n_died", 
        N = "n",
        reference = "plbo",
        time="personyears",
        family = "poisson",
        link = "cloglog",
        effects = "random")
        

}
\references{
\insertRef{gemtc}{BUGSnet}

\insertRef{TSD3}{BUGSnet}
}
