% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.regplot.R
\name{nma.regplot}
\alias{nma.regplot}
\title{Plot of relative treatment effects vs covariate values for meta-regression.}
\usage{
nma.regplot(nma, x.range = NULL, lwd = 1, palette = "Set1")
}
\arguments{
\item{nma}{Resulting output from running \code{nma.run()}.}

\item{x.range}{The range of the x axis (covariate values). By default, the range will be the same as in the data.}

\item{lwd}{Line width relative to the default (default=1).}

\item{palette}{A string indicating the colour set from RcolorBrewer for the plot. "set1" is great, but you may need a different colour set if
there are many treatments in your network.}
}
\value{
\code{regplot} - A plot of the relative treatment effects vs covariate values for meta-regression.
}
\description{
Produces a plot of relative treatment effects on the linear scale vs covariate values for meta-regression.
}
\seealso{
\code{\link{nma.run}}
}
