% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.run.auto.R
\name{nma.run.auto}
\alias{nma.run.auto}
\title{Run NMA model}
\usage{
nma.run.auto(
  model,
  monitor = "DEFAULT",
  DIC = TRUE,
  inits = "DEFAULT",
  mode = "report"
)
}
\arguments{
\item{model}{A \code{BUGSnetModel} object produced by running \code{nma.model}.}

\item{monitor}{A vector of all variables that you would like to monitor. Default is "DEFAULT" which will monitor the relative treatment effects \code{d}
as well as \code{sigma} when a random effects model is used and the regression coefficients \code{beta} when meta-regression is used.}

\item{DIC}{Default is TRUE and nodes required to calculate the DIC and other fit statistics are monitored. Otherwise you may set it to FALSE.}

\item{inits}{Specifies initial values and random number generator options for each chain. The "DEFAULT" option uses the R random seed to set the JAGS random
seeds. Non-default options are passed directly to \code{\link{jags.model}}. In order to use the JAGS default initialization, set inits to NULL.
See \code{\link{jags.model}} for more info.}

\item{mode}{String specifying what mode to use for the analysis. Choices are "quick", "report", or "paper".}
}
\value{
\code{nma.run} returns an object of class \code{BUGSnetRun} which is a list containing the following components:

\code{samples} - The MCMC samples produced by running the BUGS model.

\code{model} - The \code{BUGSnetModel} object obtained from \code{nma.model} which was used to run \code{jags}.

\code{scale} - The scale of the outcome, based on the chosen family and link function.

\code{trt.key} - Treatments mapped to numbers, used to run BUGS code.

\code{family} - Family that was used for the NMA model (e.g normal, binomial, poisson)

\code{link} - Link function that was used for the NMA model (e.g normal, binomial, poisson)
}
\description{
Takes bugs code from an object produced by \code{nma.model} and runs model using \code{jags}.
}
\seealso{
\code{\link{nma.model}}, \code{\link{nma.fit}}, \code{\link{nma.league}}, \code{\link{nma.rank}}, \code{\link{nma.forest}}, \code{\link{nma.regplot}}, \code{\link{nma.trace}}, \code{\link{jags.model}}
}
