% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.run.R
\name{nma.run}
\alias{nma.run}
\title{Run NMA model}
\usage{
nma.run(bugs, monitor = c("d"), n.adapt, n.burnin, n.iter, thin = 1,
  n.chains = 3)
}
\arguments{
\item{bugs}{Object produced by running \code{nma.model}.}

\item{monitor}{A list of all variables that you would like to monitor. The default is simply the treatment
effect samples ("d"). But you may want to monitor the deviance ("dev") as well.}
}
\value{
\code{model} - A long character string containing the model that was run in \code{rjags}.

\code{bugsdata2}

\code{scale}

\code{trt.map.table} - Treatments mapped to integer numbers, used to run BUGS code.

\code{family} - Family that was used for the NMA model (e.g normal, binomial, poisson)

\code{link} - Link function that was used for the NMA model (e.g normal, binomial, poisson)
}
\description{
Takes bugs code from an object produced by \code{nma.model} and runs model using \code{rjags}.
}
