% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.trace.R
\name{nma.trace}
\alias{nma.trace}
\title{Deprecated. Please use \code{nma.diag} instead.}
\usage{
nma.trace(nma, n = "all", thin = 1, colours = "DEFAULT")
}
\arguments{
\item{nma}{A \code{BUGSnetRun} object produced by running \code{nma.run()}.}

\item{n}{Integer which limits the number of printed variables to the first \code{n}. Default is "all" which plots every variable.}

\item{thin}{Thinning factor for the mcmc chains. Default is 1.}

\item{colours}{An optional vector of colors, one for each chain.}
}
\description{
Deprecated. Produces traceplots of the MCMC chains obtained from \code{nma.run()}
}
\seealso{
\code{\link{nma.diag}}
}
