% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pairwise}
\alias{pairwise}
\title{Pairwise meta-analysis}
\usage{
pairwise(slr, name.trt1, name.trt2, outcome, N, sd = NULL, time = NULL,
  type.outcome = "binomial", method = "MH", method.tau = "DL",
  sm = "RR")
}
\arguments{
\item{slr}{An slr data object produced by \code{data.slr()}}

\item{name.trt1}{A string indicating the name of the comparator treatment (often Placebo)}

\item{name.trt2}{A string indicating the name of the experimental treatment}

\item{outcome}{A string indicating the name of your outcome variable}

\item{N}{A string indicating the name of the variable containing the number of participants in each arm}

\item{sd}{A string (only required when type.outcome="continuous") indicating variable name
of the standard deviation of the outcome}

\item{time}{A string required when type.outcome = "rate" or "rate2". Name of variable 
indicating person-time followup (e.g person years) or study followup time}

\item{type.outcome}{A string. Options are: "binomial", "continuous", "rate" (e.g # of events and # person-years reported), 
"rate2" (e.g # events and followup time reported)}

\item{method}{A character string indicating what type of test was used. For more info, see \code{meta}'s
documentation.}

\item{method.tau}{A character string indicating which method is used to estimate the between study variance. 
Either "DL", "PM", "REML", "ML", "HS", "SJ", "HE", or "EB", can be abbreviated.}

\item{sm}{A character string indicating which summary measure ("RR", "OR", "RD", or
ASD") is to be used for pooling of studies.}
}
\value{
A forest plot as produced by the package \code{meta}

\code{raw} - dataset containing summary statistics of meta-analysis (effect estimates,
 confidence bounds, I-squared, Q-statistic)
}
\description{
implements pairwise meta-analysis via the package \code{meta}
}
\note{
Depending on the the value of \code{type.outcome}, this function will implement the functions
\code{metabin} (dichotomous outcomes), \code{metacont} (continuos outcomes), or \code{metainc} (rate outcomes)
 from the package \code{meta}.
}
\examples{
#Sample pairwise meta-analyis

#Dichotomous / binomial outcome example comparing rapid-acting inhaled insulin to placebo.
raii_vs_plbo <- pairwise(slr = diabetes.slr,
                         name.trt1 = "Placebo", 
                         name.trt2 = "raii", 
                         outcome = "n_died",
                         N = "n")
                           
# Continuous outcome example.                           
alben_vs_plbo <- pairwise(slr=weight.slr,
                         name.trt1 = "plbo",
                         name.trt2 = "alben",
                         outcome="y",
                         N="n",
                         sd="std_dev",
                         type.outcome="continuous",
                         method = "Inverse",
                         method.tau="DL",
                         sm = "SMD")

# A rate outcome example
diet1_vs_ctrl <- pairwise(slr=diet.slr,
                         name.trt1 = "control",
                         name.trt2 = "diet 1",
                         outcome="death",
                         N="n",
                         time="py",
                         type.outcome="rate",
                         method = "MH",
                         method.tau="DL",
                         sm = "IRR")
}
