% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{parkinsons_arm}
\alias{parkinsons_arm}
\title{Parkinsons arm-based data}
\format{
A dataset.
}
\usage{
data(parkinsons_arm)
}
\description{
This dataset was used to demonstrate network meta-analysis in the NICE-DSU TSD 2 (examples 5 and
8). It includes arm-based information on three out of a total of seven clinical trials examining
the effects of five different treatments for Parkinsons on the mean off-time reduction
(continuous outcome). This dataset can be used in combination with rows 9 to 15 of the
\code{parkinsons} dataset to demonstrate the use of shared parameter models in BUGSnet.
}
\examples{
data(parkinsons_arm)
}
\references{
\insertRef{TSD2}{BUGSnet}
}
\keyword{datasets}
