% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pma.R
\name{pma}
\alias{pma}
\title{Pairwise meta-analysis}
\usage{
pma(
  data,
  name.trt1,
  name.trt2,
  outcome,
  N,
  sd = NULL,
  time = NULL,
  type.outcome,
  method = "MH",
  method.tau = "DL",
  sm
)
}
\arguments{
\item{data}{A BUGSnetData object produced by \code{data.prep()}}

\item{name.trt1}{A string indicating the name of the comparator treatment (often Placebo)}

\item{name.trt2}{A string indicating the name of the experimental treatment}

\item{outcome}{A string indicating the name of your outcome variable}

\item{N}{A string indicating the name of the variable containing the number of participants in each arm}

\item{sd}{A string (only required when type.outcome="continuous") indicating variable name
of the standard deviation of the outcome}

\item{time}{A string required when type.outcome = "rate". Name of variable indicating person-time followup (e.g person years).}

\item{type.outcome}{A string. Options are: "binomial", "continuous", "rate" (e.g # of events and # person-years reported)}

\item{method}{A character string indicating what type of test was used. For more info, see \code{meta}'s
documentation.}

\item{method.tau}{A character string indicating which method is used to estimate the between study variance.
Either "DL", "PM", "REML", "ML", "HS", "SJ", "HE", or "EB", can be abbreviated.}

\item{sm}{A character string indicating which summary measure ("RR", "OR", "RD", or
ASD") is to be used for pooling of studies.}
}
\value{
A forest plot as produced by the package \code{meta}

\code{raw} - dataset containing summary statistics of meta-analysis (effect estimates,
confidence bounds, I-squared, Q-statistic)
}
\description{
implements pairwise meta-analysis via the package \code{meta}
}
\note{
Depending on the the value of \code{type.outcome}, this function will implement the functions
\code{metabin} (dichotomous outcomes), \code{metacont} (continuos outcomes), or \code{metainc} (rate outcomes)
from the package \code{meta}.
}
\examples{

data(diabetes.sim)

diabetes.slr <- data.prep(arm.data = diabetes.sim, 
varname.t = "Treatment", 
varname.s = "Study")

pma(data = diabetes.slr,
type.outcome="binomial",
sm="OR",
name.trt1 = "Placebo", 
name.trt2 = "Diuretic", 
outcome = "diabetes",
N = "n")
                           
}
\seealso{
\code{\link{data.prep}}
}
