% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sucra.table.R
\name{sucra}
\alias{sucra}
\title{SUCRA Table and Plot}
\usage{
sucra(jagsoutput, largerbetter, line.thickness = 1,
  colour.set = "set1")
}
\arguments{
\item{jagsoutput}{Resulting dataset from running \code{nma.analysis()}.}

\item{largerbetter}{A boolean variable indicating whether a larger probability should indicate a more effective treatment (TRUE) or
if a smaller probability should indicate a more effective treatment (FALSE).}

\item{line.thickness}{A number indicating the thickness of the lines in the SUCRA plot.}

\item{colour.set}{A string indicating the colour set from RcolorBrewer. "set1" is great, but may need a different colour set if 
there are many treatments in your network.}
}
\value{
\code{s.table} - A dataset containing a SUCRA table.

\code{s.plot} - A plot showing the probability of each treatment being the nth best treatment.
}
\description{
Produces a SUCRA (Surface Under the Cumulative Ranking Curve) plot and table. A Sucra table summarizes the probabilities
that each  treatment is the best, second best...worst treatment in the network.
}
\examples{

#get sucra results
sucra_results <- sucra(jagsoutput = nma_results, largerbetter = TRUE)

#plot sucra results
sucra_results$s.plot
}
