% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{thrombolytic}
\alias{thrombolytic}
\title{Thrombolytic data}
\format{A dataset.}
\usage{
data(thrombolytic)
}
\description{
Data from gemtc
}
\examples{
data(thrombolytic)

}
\keyword{datasets}
