% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{afib}
\alias{afib}
\title{Atrial Fibrillation data}
\format{
A dataset.
}
\usage{
data(afib)
}
\description{
The atrial fibrillation data was previously used to demonstrate the
network meta-analysis capabilities of the R package ‘gemtc’.
This dataset was originally generated by Cooper et al. (2009) and
was subsequently revised by van Valkenhoef et al (2013). This dataset
provides information from 25 clinical trials examining the effects of
various interventions (17) on the risk of stroke (binary outcome) among
individuals with non-rheumatic atrial fibrillation. The stroke variable is
a covariate indicating the proportion of patients in the study that had
a prior stroke.
}
\examples{
data(afib)

}
\references{
\insertRef{cooper}{BUGSnet}

\insertRef{gemtc}{BUGSnet}
}
\keyword{datasets}
