% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUGSnet-package.R
\docType{package}
\name{BUGSnet}
\alias{BUGSnet}
\alias{BUGSnet-package}
\title{BUGSnet: Bayesian network meta-analysis in compliance with best practice guidelines.}
\description{
BUGSnet (Bayesian inference Using Gibbs Sampling to conduct NETwork meta-analysis) is a feature-rich R package to conduct Bayesian network meta-analyses in compliance with best practice and reporting guidelines.
Bayesian analyses are conducted with JAGS and BUGS code is automatically generated by the package based on the user's inputs.
Outputs are highly customizable and include network plots, tables of network characteristics, league tables and league heat plots, SUCRA plots, rankograms, forest plots, leverage plots, traceplots, posterior mean deviance comparison plots.
}
\details{
To be able to use this package, you must have installed JAGS (Just Another Gibbs Sampler) on your computer.

BUGSnet is not distributed on CRAN but rather in the form of a .tar.gz source file. To install BUGSnet, run

install.packages("path/BUGSnet_1.0.1.tar.gz", repos = NULL, type = "source")

where "path/BUGSnet_1.0.1.tar.gz" should be replaced to reflect the location of the .tar.gz file on your computer.

How to cite this work? Type citation("BUGSnet").
}
\references{
\insertRef{JAGSmanual}{BUGSnet}

\insertRef{JAGSweb}{BUGSnet}
}
