% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{diabetes.sim}
\alias{diabetes.sim}
\title{Diabetes data with simulated age covariate}
\format{A dataset.}
\usage{
data(diabetes.sim)
}
\description{
Data from NICE TSD 2.
}
\examples{
data(diabetes.sim)

}
\keyword{datasets}
