% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{diabetes.sim}
\alias{diabetes.sim}
\title{Diabetes data with simulated age covariate}
\format{
A dataset.
}
\usage{
data(diabetes.sim)
}
\description{
This is the same as the Diabetes dataset but we have
added a simulated age variable to this dataset for the purpose of
demonstrating certain features within BUGSNET.
}
\examples{
data(diabetes.sim)

}
\references{
\insertRef{TSD2}{BUGSnet}

\insertRef{elliott}{BUGSnet}
}
\keyword{datasets}
