% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.plot.R
\name{net.plot}
\alias{net.plot}
\title{Network Plot}
\usage{
net.plot(
  data,
  node.scale = 5,
  edge.scale = 2,
  flag = NULL,
  study.counts = FALSE,
  label.offset1 = 0,
  label.offset2 = 1,
  graph.scale = TRUE,
  node.lab.cex = 1,
  edge.lab.cex = 1,
  node.colour = "#f69c54",
  edge.colour = "grey",
  edge.lab.colour = "blue",
  flag.edge.colour = "lightpink",
  layout = "layout_in_circle",
  layout.params = NULL
)
}
\arguments{
\item{data}{A \code{BUGSnetData} object produced by \code{data.prep()}.}

\item{node.scale}{Size of the nodes (default=5)}

\item{edge.scale}{Thickness of the edges (default=2).}

\item{flag}{Used to highlight direct comparisons to particular treatments (optional).
Set this value to treatment(s) of interest and it will highlight, in red, all of the edges
going into the specified treatment(s).}

\item{study.counts}{If TRUE, prints the number of studies on each edge.}

\item{label.offset1}{Node label location (x-axis) relative to node. Default=0}

\item{label.offset2}{Node label location (y-axis) relative to node. Default=1}

\item{graph.scale}{Whether to make edges and nodes proportionnaly larger with the number of studies/arms. Default is TRUE.}

\item{node.lab.cex}{Size of node labels}

\item{edge.lab.cex}{Size of edge labels}

\item{node.colour}{Node colour (string)}

\item{edge.colour}{Edge colour (string)}

\item{edge.lab.colour}{Edge label colour (string)}

\item{flag.edge.colour}{Color of flagged edges (string)}

\item{layout}{Specifies how the nodes should be layed out in the graph. Default is "layout_in_circle". See \code{igraph::layout_}
for layout options}

\item{layout.params}{Additional parameters to be passed to the chosen 'layout' function. See \code{igraph::layout_} for more information}
}
\description{
Produces network plot where nodes represent treatments and edges represent direct
evidence (e.g an RCT) comparing treatments.
}
\examples{

data(diabetes.sim)

diabetes.slr <- data.prep(arm.data = diabetes.sim, 
varname.t = "Treatment", 
varname.s = "Study")

# use default settings
net.plot(diabetes.slr)

# Highlight all direct comparisons with Placebo. Adjust node and edge size, centre node labels
net.plot(diabetes.slr, 
node.scale=4, 
edge.scale=1.5, 
flag="Placebo", 
label.offset1=0, 
label.offset2=0)
}
\seealso{
\code{\link{data.prep}}, \code{igraph::layout_}
}
