% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.fit.R
\name{nma.compare}
\alias{nma.compare}
\title{Consistency vs Inconsistency plot}
\usage{
nma.compare(consistency.model.fit, inconsistency.model.fit, ...)
}
\arguments{
\item{consistency.model.fit}{Results of \code{nma.fit()} of an consistency model.}

\item{inconsistency.model.fit}{Results of \code{nma.fit()} of an inconsistency model.}

\item{...}{Graphical arguments such as main=, ylab=, and xlab= may be passed in \code{plot()}.}
}
\description{
Plots the posterior mean deviance of a consistency model vs an inconsistency model. This plot can help identify loops
where inconsistency is present. Ideally, both models will contribute approximately 1 to the posterior mean deviance.
}
\examples{
# Assess model fit for a both an inconsistency model and consistency model using nma.fit()
assess.consistency <- nma.fit(consistency_results)
assess.inconsistency <- nma.fit(inconsistency_results)

#Plot the results against each other to assess inconsistency
inconsistency.plot(consistency_results, inconsistency_results)
}
