% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.fit.R
\name{nma.fit}
\alias{nma.fit}
\title{Assess Model Fit}
\usage{
nma.fit(nma, plot.pD = TRUE, plot.DIC = TRUE, plot.Dres = TRUE, ...)
}
\arguments{
\item{nma}{Resulting output from running \code{nma.analysis()}.}

\item{...}{Graphical arguments such as main=, ylab=, and xlab= may be passed as in \code{plot()}. These arguments will only effect the
leverage plot.}
}
\value{
\code{DIC} - A number indicating the Deviance Information Criteria. The DIC is calculated as the sum of \code{Dres} and \code{pD}.
A larger DIC is indicative of a worse model fit.

\code{leverage} - A vector with one value per study arm indicating the leverage of each data point (study arm). Leverage is defined as \code{pmdev} minus the
deviance at the posterior mean of the fitted values.

\code{w} - A vector with one value per study arm. The magnitude of \code{w} represents the data point's contribution to the posterior mean deviance of the
model and is simply the square root of \code{pmdev}. The sign indicates whether the data point is being over (negative sign) or under (positive sign) estimated by
the model and is calculated as the sign of the difference of the observed outcome minus the predicted outcome.

\code{pmdev} - A vector with one value per study arm representing the posterior mean residual deviance for each data point (study arm).

\code{Dres} - The posterior mean of the residual deviance.

\code{pD} - The effective number of parameters, calculated as the sum of the leverages.
}
\description{
Computes the Deviance Information Criteria (DIC) and produces a leverage plot (as in the NICE Technical Support Document 2)
for a given model. These can be used to assess and compare the fit of different models (i.e fixed vs random effects, consistency vs
inconsistency). \code{nma.fit} also produces a plot comparing the leverage of each data point against their contribution to
the total posterior deviance. Any point lying outside the purple dotted line is considered to be an "outlier";
contributing to the model's poor fit.
}
\examples{
#Compare fixed vs random effects via leverage plots and DIC
#fixed_effects_results are the outputs of nma.analysis() where the parameter effects="fixed"
#random_effects_results are the outputs of nma.analysis() where the parameter effects="random"

par(mfrow=c(1,2))
nma.fit(fixed_effects_results, main = "Fixed Effects Model" )
nma.fit(random_effects_results, main= "Random Effects Model")
}
