% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.league.R
\name{nma.league}
\alias{nma.league}
\title{League Table and Heat Plot}
\usage{
nma.league(
  nma,
  central.tdcy = "median",
  log.scale = FALSE,
  order = NULL,
  low.colour = "darkgoldenrod1",
  mid.colour = "white",
  high.colour = "cornflowerblue",
  cov.value = NULL,
  digits = 2
)
}
\arguments{
\item{nma}{A \code{BUGSnetRun} object produced by running \code{nma.run()}.}

\item{central.tdcy}{The statistic that you want to use in order to measure relative effectiveness. The options are "mean" and "median".}

\item{log.scale}{If TRUE, odds ratios, relative risk or hazard ratios are reported on the log scale. Default is FALSE.}

\item{order}{A vector of strings representing the order in which to display the treatments.}

\item{low.colour}{A string indicating the colour of low relative treatment effects for the heat plot (e.g relative risk of 0.5).}

\item{mid.colour}{A string indicating the colour of null relative treatment effects for the heat plot (e.g relative risk of ~1.0).}

\item{high.colour}{A string indicating the colour of high relative treatment effects for the heat plot (e.g relative risk of ~2.0).}

\item{cov.value}{Must be specified for meta-regression. This is the value of the covariate for which to report the results.}

\item{digits}{Number of digits to display after the decimal point}
}
\value{
\code{table} - A league table. Row names indicate comparator treatments.

\code{longtable} - League table in the long format.

\code{heatplot} - League heat plot, where a color scale is used to represent relative treatment effects and ** are used to highlight statistically significant differences.
}
\description{
Produces a league table and a league heat plot that contain point estimates of relative effectiveness
for all possible pairs of treatments point estimates along with 95\% credible intervals obtained with the quantile method.
}
\examples{
data(diabetes.sim)

diabetes.slr <- data.prep(arm.data = diabetes.sim, 
varname.t = "Treatment", 
varname.s = "Study")

#Random effects, consistency model.
#Binomial family, cloglog link. This implies that the scale will be the Hazard Ratio.
diabetes.re.c <- nma.model(data = diabetes.slr,
       outcome = "diabetes", 
       N = "n",
       reference = "Placebo",
       family = "binomial",
       link = "cloglog",
       effects = "random",
       type="consistency",
       time="followup"
       )
 
diabetes.re.c.res <- nma.run(diabetes.re.c,
n.adapt=100,
n.burnin=0,
n.iter=100)

 
league_table <- nma.league(nma=diabetes.re.c.res, central.tdcy="median")
league_table$heatplot
league_table$table
league_table$longtable
}
\seealso{
\code{\link{nma.run}}, \code{\link{nma.rank}}, \code{\link{nma.forest}}
}
