% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.rank.R
\name{nma.rank}
\alias{nma.rank}
\title{SUCRA Table and Plot}
\usage{
nma.rank(nma, largerbetter, lwd = 1, colour.set = "Set1")
}
\arguments{
\item{nma}{Resulting dataset from running \code{nma.run()}.}

\item{largerbetter}{A boolean variable indicating whether a larger probability should indicate a more effective treatment (TRUE) or
if a smaller probability should indicate a more effective treatment (FALSE).}

\item{lwd}{Line width relative to the default (default=1) in the SUCRA plot.}

\item{colour.set}{A string indicating the colour set from RcolorBrewer. "set1" is great, but you may need a different colour set if 
there are many treatments in your network.}
}
\value{
\code{table} - A rank table as a dataset object.

\code{sucra} - A SUCRA plot showing the probability of each treatment being the nth best treatment.

\code{rankogram} - A rankogram plot showing the probability of each treatment being the nth best treatment.

\code{ranks} - A vector containing the order of efficacy of treatments (from best to worst). This value 
is useful when creating the league heat plot with \code{nma.league()}.
}
\description{
Produces a SUCRA (Surface Under the Cumulative Ranking Curve) plot and table. A Sucra table summarizes the probabilities
that each  treatment is the best, second best...worst treatment in the network.
}
\examples{

#get sucra results
sucra_results <- nma.rank(nma = nma_results, largerbetter = TRUE)

#plot sucra results
sucra_results$sucra
}
