% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.rank.R
\name{nma.rank}
\alias{nma.rank}
\title{Table and Plots of Treatment Rankings}
\usage{
nma.rank(
  nma,
  largerbetter,
  sucra.lwd = 1,
  sucra.palette = "Set1",
  ranko.palette = "Blues",
  cov.value = NULL
)
}
\arguments{
\item{nma}{A \code{BUGSnetRun} object produced by running \code{nma.run()}.}

\item{largerbetter}{A boolean variable indicating whether a larger probability should indicate a more effective treatment (TRUE) or
if a smaller probability should indicate a more effective treatment (FALSE).}

\item{sucra.lwd}{Line width relative to the default (default=1) in the SUCRA plot.}

\item{sucra.palette}{A string indicating the colour set from RcolorBrewer for the SUCRA plot. "Set1" is used by default and is automatically extended if
there are many treatments in your network.}

\item{ranko.palette}{A string indicating the colour set from RcolorBrewer for the rankogram. "Blues" is used by default and is automatically extended if
there are many treatments in your network.}

\item{cov.value}{Must be specified for meta-regression. This is the value of the covariate for which to report the results.}
}
\value{
\code{ranktable} - A rank table showing the probability of each treatment being the nth best treatment.

\code{sucratable} - A table showing the probability of each treatment being the nth best treatment or better and an overall SUCRA value for each treatment.

\code{order} - A vector containing the order of efficacy of treatments (from best to worst) based on their SUCRA value. This vector
is useful for ordering treatments when creating the league heat plot with \code{nma.league()}.

\code{longtable} - A long form table of ranking probabilities and SUCRA value.

\code{sucraplot} - A SUCRA plot showing the probability of each treatment being the nth best treatment or better.

\code{rankogram} - A rankogram showing the probability of each treatment being the nth best treatment.
}
\description{
Produces a SUCRA (Surface Under the Cumulative Ranking Curve) plot and table. A Sucra table summarizes the probabilities
that each treatment is the best, second best...worst treatment in the network.
}
\examples{
data(diabetes.sim)

diabetes.slr <- data.prep(arm.data = diabetes.sim, 
varname.t = "Treatment", 
varname.s = "Study")

#Random effects, consistency model.
#Binomial family, cloglog link. This implies that the scale will be the Hazard Ratio.
diabetes.re.c <- nma.model(data = diabetes.slr,
       outcome = "diabetes", 
       N = "n",
       reference = "Placebo",
       family = "binomial",
       link = "cloglog",
       effects = "random",
       type="consistency",
       time="followup"
       )
 
diabetes.re.c.res <- nma.run(diabetes.re.c,
n.adapt=100,
n.burnin=0,
n.iter=100)
 
#get sucra results
sucra_results <- nma.rank(nma = diabetes.re.c.res, largerbetter = FALSE)

#plot sucra results
sucra_results$sucraplot
}
\seealso{
\code{\link{nma.run}}, \code{\link{nma.league}}, \code{\link{nma.forest}}
}
