% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.run.R
\name{nma.run}
\alias{nma.run}
\title{Run NMA model}
\usage{
nma.run(
  model,
  monitor = "DEFAULT",
  DIC = TRUE,
  n.adapt = 1000,
  n.burnin = floor(n.iter/2),
  n.iter,
  thin = 1,
  n.chains = 3,
  inits = "DEFAULT"
)
}
\arguments{
\item{model}{A \code{BUGSnetModel} object produced by running \code{nma.model}.}

\item{monitor}{A vector of all variables that you would like to monitor. Default is "DEFAULT" which will monitor the relative treatment effects \code{d}
as well as \code{sigma} when a random effects model is used and the regression coefficients \code{beta} when meta-regression is used.}

\item{DIC}{Default is TRUE and nodes required to calculate the DIC and other fit statistics are monitored. Otherwise you may set it to FALSE.}

\item{n.adapt}{Number of adaptations for the mcmc chains.}

\item{n.burnin}{Number of burnin iterations for the mcmc chains.}

\item{n.iter}{Number of iterations for the mcmc chains.}

\item{thin}{Thinning factor for the mcmc chains. Default is 1.}

\item{n.chains}{Number of mcmc chains. Default is 3.}

\item{inits}{Specifies initial values and random number generator options for each chain. The "DEFAULT" option uses the R random seed to set the JAGS random
seeds. Non-default options are passed directly to \code{\link{jags.model}}. In order to use the JAGS default initialization, set inits to NULL.
See \code{\link{jags.model}} for more info.}
}
\value{
\code{nma.run} returns an object of class \code{BUGSnetRun} which is a list containing the following components:

\code{samples} - The MCMC samples produced by running the BUGS model.

\code{model} - The \code{BUGSnetModel} object obtained from \code{nma.model} which was used to run \code{jags}.

\code{scale} - The scale of the outcome, based on the chosen family and link function.

\code{trt.key} - Treatments mapped to numbers, used to run BUGS code.

\code{family} - Family that was used for the NMA model (e.g normal, binomial, poisson)

\code{link} - Link function that was used for the NMA model (e.g normal, binomial, poisson)
}
\description{
Takes bugs code from an object produced by \code{nma.model} and runs model using \code{jags}.
}
\examples{
data(diabetes.sim)

diabetes.slr <- data.prep(arm.data = diabetes.sim, 
varname.t = "Treatment", 
varname.s = "Study")

#Random effects, consistency model.
#Binomial family, cloglog link. This implies that the scale will be the Hazard Ratio.
diabetes.re.c <- nma.model(data = diabetes.slr,
       outcome = "diabetes", 
       N = "n",
       reference = "Placebo",
       family = "binomial",
       link = "cloglog",
       effects = "random",
       type="consistency",
       time="followup"
       )
 
diabetes.re.c.res <- nma.run(diabetes.re.c,
n.adapt=100,
n.burnin=0,
n.iter=100)
       
}
\seealso{
\code{\link{nma.model}}, \code{\link{nma.fit}}, \code{\link{nma.league}}, \code{\link{nma.rank}}, \code{\link{nma.forest}}, \code{\link{nma.regplot}}, \code{\link{nma.trace}}, \code{\link{jags.model}}
}
