% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.trace.R
\name{nma.trace}
\alias{nma.trace}
\title{Traceplot of MCMC chains}
\usage{
nma.trace(jagsoutput, n = "all")
}
\arguments{
\item{jagsoutput}{An output produced by \code{nma.run()}}

\item{n}{}
}
\description{
Produces traceplots of the MCMC chains obtained from \code{nma.run()}
}
