% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.trace.R
\name{nma.trace}
\alias{nma.trace}
\title{Traceplot of MCMC chains}
\usage{
nma.trace(nma, n = "all", thin = 1, colours = "DEFAULT")
}
\arguments{
\item{nma}{A \code{BUGSnetRun} object produced by running \code{nma.run()}.}

\item{n}{Integer which limits the number of printed variables to the first \code{n}. Default is "all" which plots every variable.}

\item{thin}{Thinning factor for the mcmc chains. Default is 1.}

\item{colours}{An optional vector of colors, one for each chain.}
}
\description{
Produces traceplots of the MCMC chains obtained from \code{nma.run()}
}
\examples{
data(diabetes.sim)

diabetes.slr <- data.prep(arm.data = diabetes.sim, 
varname.t = "Treatment", 
varname.s = "Study")

#Random effects, consistency model.
#Binomial family, cloglog link. This implies that the scale will be the Hazard Ratio.
diabetes.re.c <- nma.model(data = diabetes.slr,
       outcome = "diabetes", 
       N = "n",
       reference = "Placebo",
       family = "binomial",
       link = "cloglog",
       effects = "random",
       type="consistency",
       time="followup"
       )
 
diabetes.re.c.res <- nma.run(diabetes.re.c,
n.adapt=1000,
n.burnin=1000,
n.iter=10000)

nma.trace(diabetes.re.c.res, n=4, thin=10)
 
}
\seealso{
\code{\link{nma.run}}
}
