% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.trace.R
\name{nma.trace}
\alias{nma.trace}
\title{Traceplot of MCMC chains}
\usage{
nma.trace(nma, n = "all")
}
\arguments{
\item{nma}{An output produced by \code{nma.run()}}

\item{n}{Limits the number of printed variables to the first \code{n}}
}
\description{
Produces traceplots of the MCMC chains obtained from \code{nma.run()}
}
\seealso{
\code{\link{nma.run}}
}
