% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{parkinsons}
\alias{parkinsons}
\title{Parkinsons contrast data}
\format{
A dataset.
}
\usage{
data(parkinsons)
}
\description{
This dataset was used to demonstrate network meta-analysis in the NICE-DSU TSD 2 (examples 5, 7, and 8). It includes information on seven clinical trials examining
the effects of five different treatments for Parkinsons on the mean off-time reduction (continuous outcome).
This dataset is in the form of mean differences in order to demonstrate the contrast data.
}
\examples{
data(contrast)
\insertRef{TSD2}{BUGSnet}

}
\references{
\insertRef{walton}{BUGSnet}
}
\keyword{datasets}
